add ms_exec_binary libms

+++ libms/get_next_line.c
@@ -6,11 +6,11 @@
 /*   By: DERYCKE <DERYCKE@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/15 15:47:16 by dideryck          #+#    #+#             */
-/*   Updated: 2018/07/20 22:16:48 by DERYCKE          ###   ########.fr       */
+/*   Updated: 2018/08/23 22:21:44 by DERYCKE          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/get_next_line.h"
+#include "get_next_line.h"
 
 static void		create_line(char **rest, char **line)
 {
@@ -47,7 +47,7 @@ static void		add_buffer_to_rest(char **rest, char buffer[], int ret)
 		*rest = ft_strjoin_free(*rest, buffer);
 }
 
-int				get_next_line(const int fd, char **line)
+ssize_t				get_next_line(const int fd, char **line)
 {
 	static t_pos		pos;
